I've been working on the user interface to Sire. I have decided to use a series of XML files to describe the various options available to a simulation. These describe the options, which are represented in C++ as Option and Value objects. These can read and write themselves to simple text files, thereby providing a friendly, high-level consistent interface to the Sire-derived programs. In addition, the XML files make it easy for other programs to read and write Sire config files, thereby easing the integration of Sire into other packages and GUIs.
See share/SireSim/proteinligandmc.xml for an idea of what the XML files will look like,
though bear in mind this is still a work in progress ;-)

Code compiles, links and appears to work well. Python wrappers should not be auto-generated
for this version, as there are lots of missing functions and you will get link errors
when you try to link the resulting wrappers.