I've separated out the SireSim options library into a new library called "Conspire".
In this, I have put a cut-down version of Siren, so that Conspire can either compile
as a standalone project, or it can link against the larger, standard Siren library.

Code compiles, links and runs as expected, although many unit tests need to be
written to make sure that the port from Sire to Siren is correct.