I've debugged and fixed problems in IntraCLJPotential and IntraSoftCLJPotential and have checked that the energies given by both agree with one another, and that having different coulomb and LJ cutoffs gives the same numbers as forcefields with same cutoffs (see test/SireMM/testintrasoftcljff.py). I will next work on a similar test for InterCLJPotential and InterSoftCLJPotential.
Note I also found a bug in SireFF/detail/ffmolecules.cpp that meant that it crashed
when calculating intramolecular energies of molecules when the molecule had been
added to the forcefield one residue at a time. The idx_to_cgidx map was not being
updated when the atom selection was being changed.