I've fixed a lot of performance problems with MultiFloat and MultiDouble that were revealed when I updated the speedtest to do some realistic calculation (coulomb energy between two groups of atoms). I've inlined most things and have also added SP/DP sums for the floating point summation (which gives almost exactly the same result as for DP/DP).
The speed results are in speedtest_results, and show that the AVX floating point code
is going as fast as the hardware will allow (13.8-13.9 GFLOPs per core, which is almost
exactly the theoretical peak of 14.4 GFLOPs per core...! 95% efficient).

The speedtest also shows that all versions (vanilla, SSE and AVX, for float and double)
give the same energy, which is reassuring.