Added mixed precision OpenCL mode by default in openmmmdintegrator as this appears to be the only way to use the platform with the Intel OpenCL SDK
Also changed the free energy code to set constraints between bonds that are in a perturbed solute, if the ff parameters do not change. 

The constraint code probably needs to be revised. It should be possible to constrain perturbed bonds to their target equilibrium value and get 
reasonabe free energies without having to use a small timestep.