I've moved the energy code into a special function and created a CLJAtoms object that holds just the vectorised coordinates and parameters of a set of atoms. I will use this, plus energy calculating functions as the basis of the code to calculate fast CLJ energies. Timing shows that the code is now accurate and very quick (see test/SireMM/testgridff2.py), with most of the time (60%) now being the conversion from the current memory layout of a CoordGroup into the CLJAtoms layout. I will need to update GridFF2 to store the atoms using the new layout and will need to work out the best way of maintaining this through the simulation...
Also, I found that adding "if" statements inside the energy loop was really painful,
even if they were hard coded to give the same answer (e.g. always use arithmetic
combining rules). I guess that this prevents pipelining of the code, and so
I will have to go down the route of creating specialised CLJ energy functions for
each type of calculation... (arithmetic, geometric, periodic box, cartesian,
soft-core, each different cutoff type...!). I've also decided that I will abandon
group-based cutoffs as they are a pain in the neck to code and have been shown
to give a poor representation of the electrostatics.

Finally, I've seen that the grid evaluation code is so fast that it has negligable
impact on the time.