A lot of changes so that Class::update(molecule) or Class::update(molecules) changes to Class::update(molecule, bool auto_update=true) and Class::update(molecules, bool auto_update=true).
This will ensure that the default behaviour of the old code (no partially-updated groups
and forcefields) is preserved for older code that is unaware of the time-savings 
(and danger) of partial updates. Now, the user will have to explicitly ask for partial
updating, and so should be aware of the limitations.

All compiles but need to generate python wrappers