Found the cause of exceptions causing a segfault - it seems to be that -fomit-frame-pointer causes problems on a mac when the exception moves from C++ to python. I've removed this option from the CMakeFiles.
I have also turned back on backtraces and added code to exception so that
'slow' exceptions can be disabled. This allows code that throws lots
of exceptions (correctly, and are handled correctly) to run quickly
without being slowed down by the need to generate backtraces or other
human-useful information

Found the cause of the segfault in the test_interff_residue.py unit
test. The segfault was because of an unhandled exception. The exception
is because something is going wrong in CLJExtractor::_pvt_update when
part of a multi-residue molecule is changed.

Code compiles, links and runs, but the test_interff_residue.py
unit test fails